split examples service discoverer
