+++ examples/netapp_capif_connector_examples.py
@@ -26,44 +26,8 @@ def showcase_capif_connector():
 
     capif_connector.register_and_onboard_netapp()
 
-def showcase_service_discovery():
-    service_discoverer = ServiceDiscoverer(folder_path_for_certificates_and_api_key="/home/alex/Projects/test_certificate_folder",
-                                           capif_host="capifcore",
-                                           capif_https_port=443
-                                           )
-    endpoints = service_discoverer.discover_service_apis()
-    print(endpoints)
-
-def showcase_retrieve_endpoint_url():
-    service_discoverer = ServiceDiscoverer(folder_path_for_certificates_and_api_key="/home/alex/Projects/test_certificate_folder",
-                                           capif_host="capifcore",
-                                           capif_https_port=443
-                                           )
-    url = service_discoverer.retrieve_specific_resource_name(
-        "/nef/api/v1/3gpp-monitoring-event/",
-        "MONITORING_SUBSCRIPTIONS"
-    )
-    print(url)
-
-
-def showcase_access_token_retrieval_from_capif():
-    service_discoverer = ServiceDiscoverer(folder_path_for_certificates_and_api_key="/home/alex/Projects/test_certificate_folder",
-                                           capif_host="capifcore",
-                                           capif_https_port=443
-                                           )
-    endpoints = service_discoverer.discover_service_apis()
-    if len(endpoints)>0:
-        api_name = endpoints["serviceAPIDescriptions"][0]["apiName"]
-        api_id =  endpoints["serviceAPIDescriptions"][0]["apiId"]
-        aef_id =  endpoints["serviceAPIDescriptions"][0]["aefProfiles"][0]["aefId"]
-        access_token = service_discoverer.get_access_token(api_name,api_id,aef_id)
-        print(access_token)
-    else:
-        print("no endpoints have been registered. Make sure NEF has registered to CAPIF first")
 
 if __name__ == "__main__":
     #Let's register NetApp to CAPIF. This should happen exactly once
     showcase_capif_connector()
-    showcase_service_discovery()
-    showcase_retrieve_endpoint_url()
-    showcase_access_token_retrieval_from_capif()
+
